pyproject.toml
[project]
name = "supplychainrag"
version = "0.1.0"
description = "Supply Chain AI Assistant with Governable RAG and Progressive Disclosure"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "faiss-cpu>=1.7.4",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.2.2",
    "openai>=1.0.0",
    "docker>=7.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "rich>=13.0.0",
    "jinja2>=3.1.2",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "jieba>=0.42.1",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

# Use setuptools instead of hatchling for simpler configuration
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
